+++ dart/geobase/lib/coordinates.dart
@@ -27,6 +27,7 @@ export 'src/coordinates/base/position.dart';
 export 'src/coordinates/base/position_extensions.dart';
 export 'src/coordinates/base/position_series.dart';
 export 'src/coordinates/base/positionable.dart';
+export 'src/coordinates/base/value_positionable.dart';
 export 'src/coordinates/geographic/dms.dart';
 export 'src/coordinates/geographic/geobox.dart';
 export 'src/coordinates/geographic/geographic.dart';